wip: develop for rmuc

+++ src/cxx/context.cc
@@ -57,8 +57,8 @@ struct Context::Impl {
             /*^^*/ if constexpr (std::is_enum_v<T>) {
                 using U = std::underlying_type_t<T>;
                 to_sync = static_cast<T>(data.as<U>());
-            } else if constexpr (std::is_constructible_v<T, std::uint8_t>) {
-                to_sync = T{data.as<std::uint8_t>()};
+            } else if constexpr (std::is_constructible_v<T, std::uint16_t>) {
+                to_sync = T{data.as<std::uint16_t>()};
             } else {
                 to_sync = data.as<T>();
             }
